// apps/humayro/lib/api/model/pageResponseBranchDTO.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Humayro API
 * Humayro e-commerce API
 * OpenAPI spec version: 1.0
 */
import type { BranchDTO } from './branchDTO';

export interface PageResponseBranchDTO {
  content?: BranchDTO[];
  page?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
  last?: boolean;
}