// apps/humayro/lib/api/model/pageResponseDistrictDTO.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Humayro API
 * Humayro e-commerce API
 * OpenAPI spec version: 1.0
 */
import type { DistrictDTO } from './districtDTO';

export interface PageResponseDistrictDTO {
  content?: DistrictDTO[];
  page?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
  last?: boolean;
}