// apps/humayro/lib/api/model/pageResponseRegionDTO.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Humayro API
 * Humayro e-commerce API
 * OpenAPI spec version: 1.0
 */
import type { RegionDTO } from './regionDTO';

export interface PageResponseRegionDTO {
  content?: RegionDTO[];
  page?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
  last?: boolean;
}